optionally log timing of requests
Timing logging is disabled by default. To enable timing set the environment
variable `NODE_REQUEST_LOG_TIMING` to `true`. Set `NODE_REQUEST_LOG_TIMING`
to `false` to explicitly disable timing logging.

When enabled the timing of request events is written to stderr as JSON
of the form:

```
{
  "facility": "request",
  "method": "GET",
  "uri": "http://exmaple.com/some-resource",
  "event":"end",
  "time":172
}
```

Events of interest such as starting, ending, and aborting a request are
logged. This works well for simple requests that have only a start and
end. No attempt was made to handle more interesting requests that stream
so they may not produce useful timing logs.